cron: refactor candidate files to hard delete logic
was only being used in this script so made sense to remove from s3_client
simplifies tests a little too, and made adding retries for connection issues easier.
